Fix the issue with crashing when enabling coin
The problem was in the concurrent execution of methods. It starts to update wallets list in one thread. Then it updates the balance for newly enabled coin in the second thread before the updating wallet list is finished. Since the wallet list doesn't contain the enabled coin it doesn't find it and crashes.