Prevent scrolling back and forth when restoring state of filter coin tab
When liveData observed as state for composable it emits initial value and then the value of liveData itself. It's true also for case when restoring the view. It results in scrolling to the start and then scrolling to the last position. To prevent this behavior filterCoinsLiveData observed without default value.