Fix the issue with crash on process restore
There is a case when the app process is killed and restored by Android OS. When the app is restored it tries to restore all fragments. But the navController is not getting restored and it contains single back stack entry, which is mainFragment.

So if user was in the page like TokenBalanceFragment or TransactionInfoFragment, those fragments are restored. Since they do not get expected data, they call `navController.popBackStack()`. It results in popping off the single mainFragment of the back stack. And the MainFragment tries to create TransactionsViewModel using the mainFragment back stack entry. But the nav controller back stack is empty. That is why it crashes app.

Fixed issue by ensuring that those fragments closes themselves if needed.